Made penalties work Server GUI
